Remove MechSound class
