+++ VisualPinball.Unity/VisualPinball.Unity.Editor/Sound/SoundComponentInspector.uxml.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5edbe5d4628bdb545a05bfa87ed2f700
+guid: aad56dcd329ccee4f901eda740ace64d
 ScriptedImporter:
   internalIDToNameTable: []
   externalObjects: {}
