--- VisualPinball.Unity/VisualPinball.Unity.Editor/Sound/SoundsComponentInspector.uss
@@ -1,5 +1,0 @@
-.custom-label {
-    font-size: 20px;
-    -unity-font-style: bold;
-    color: rgb(68, 138, 255);
-}
