--- VisualPinball.Unity/VisualPinball.Unity.Editor/Sound/SoundsComponentInspector.uss.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 30f22b4450bbae84eb18d9bc7f656e8c
-ScriptedImporter:
-  internalIDToNameTable: []
-  externalObjects: {}
-  serializedVersion: 2
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
-  script: {fileID: 12385, guid: 0000000000000000e000000000000000, type: 0}
-  disableValidation: 0
