--- VisualPinball.Unity/VisualPinball.Unity.Editor/Sound/SoundsComponentInspector.uxml.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: a6e8546981c540948ae9851157602432
-ScriptedImporter:
-  internalIDToNameTable: []
-  externalObjects: {}
-  serializedVersion: 2
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
-  script: {fileID: 13804, guid: 0000000000000000e000000000000000, type: 0}
