--- VisualPinball.Unity/VisualPinball.Unity/Sound/MechSound.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: e750cca599b34fd1aa5c29d5b38788f4
-timeCreated: 1677682479
