+++ VisualPinball.Unity/VisualPinball.Unity/Sound/SoundComponent.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c28cae51000318145b40983f440013ab
+guid: 5202e8a88536f8243974d06c584c6be4
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
